Fixes #65, implements autocomplete for adding tags via the admin interface.
This uses the Scriptaculous Autocompleter library and adds a corresponding controller action and table method for data retrieval.  Autocomplete for tags now works on the items/show and items form pages.  Autocomplete will search through all existing tag names regardless of whether they are for Exhibits or Items, and it will just search alphabetically.

The Scriptaculous implementation requires that an HTML unordered list be returned from the server for every request, so a special TagsController::autocompleteAction() is available for that purpose.  The 'tokens' attribute in the Javascript ensures that autocomplete will append new tags instead of replacing the existing tags in the list.