Fixes a potential security hole involving default JSON outputs for controllers.
This removes the default action contexts from Omeka_Controller_Action and adds them to each relevant controller.  The default action contexts were 'json', 'rss2' and 'dc' for 'browse' actions, and 'json' and 'dc' for 'show' actions.  This not make sense for most controllers, since RSS and Dublin Core are only for the ItemsController.  It also opened up a security hole where sensitive data could have been passed to an attacker through the automatically serialized JSONP output.

Note that this breaks backwards compatibility for any plugin controllers that extend off Omeka_Controller_Action, but only if those controllers rely on the default existence of the 'json' output for 'browse' and 'show' actions.