Fixes #647, Item view helper needs to be able to be called inside filters.
This refactors the RecordMetadata view helper base class so that it does not maintain any state between calls.

What this means is that item() can be called inside a filter that normally runs inside of another call to item(), so essentially a plugin writer (or theme writer) can examine the value of an element while filtering another element.