Refs #154, adds an 'initializeSession' phase to Omeka_Core.
This also adds a 'session.name' field to the config.ini file.  If this is value is given, it will be used for the name of the session.  Otherwise the session name will be automatically generated from the WEB_ROOT of the installation.

The session name can be customized so that session data could potentially be shared across installations or with other applications (but not by default).

The previous behavior was that the session name was not configured in Omeka, which led to overlap between auth session namespaces on the same domain.  This was solved by using an arbitrary auth namespace that was unique to each installation (at one point stored in the database, then just using the root web URL).