Refactors detachCollectorsFromEntities migration so that all table names are referenced using the $db->ModelName syntax.
This avoids hardcoding the table name and thus prevents errors when the table
name or prefix has been changed at runtime for a given model.

Note that this behavior may result in an error if the model cannot be found, i.e.
the model has been removed in a future version of the code.  This can be solved
by temporarily injecting the missing table name into Omeka_Db before running the
migrations.