Fixes #993 and #1011, adds methods for random items and derivative images.
Adds a filter to the ItemTable for returning items with derivative images, and
adds a function for ordering results randomly. Both have been added to
applySearchFilters, called using 'hasImage' and 'random', respectively.
Refactors findRandomFeaturedItem to use applySearchFilters with parameters set
for 'random', 'featured, and 'hasImage'.

Adds new test files for the ItemTable and the get_items() helper.