Get rid of weird error actions on all controllers.
The error and forbidden actions are practically unused. Almost
everywhere just throws 404/403 exceptions or directly forwards to the
error controller.

The "404" and "403" URLs are similarly useless, but at
least the ExhibitBuilder and SimplePages currently redirect to them
sometimes to raise an error page, so they're simply re-routed more
directly to where they were going anyway.