Fix RecordMetadata to work with non-Element data
The big change here is that, internally, RecordMetadata no longer keys
off of whether or not what it got back was an array. Previously, it
assumed that any "property" would never be an array, so it called
ElementText methods on any elements of a returned array. This was a
barrier to using this helper to replace collection(), since collectors
is an array property of a Collection.

Now, RecordMetadata simply checks to see if the "text" it's dealing with
is actually an ElementText record. Aside from enabling array properties,
this has the happy side-effect of making the code a lot more readable
and less crazy. This also made it easier to apply the normal "index"
filter up-front, so we no longer waste time filtering and processing
texts that we're going to throw away.