Delete most of the index.html files.
Exceptions are the ones in the archive subdirectories, since we need at
least one file in each for Git to track them, and the install
directory's index.html, which actually serves a purpose.

The rest of them are just a holdover from when our rewrite config
didn't pick up URLs that point to real directories.