Add TravisCI Support
TravisCI is a hosted continuous integration service integrated with github.
TravisCI is automated to run tests when a committer pushes to Github, and
allows feedback if someone breaks the build. This also allows the testing of
pull requests from github, so patches don't also break the build.

There's also a nice banner image you can add to the README that displays the
current build status for the project.