Add admin bar to public site for logged-in users.
Adds a 'common/admin-bar.php' partial for admin bar markup. Includes a
'admin_bar_nav' filter for modifying the contents of the admin bar navigation.

Admin bar is styled with 'css/admin-bar.css'. It is absolutely
positioned at the top of the the theme. The Google Fonts CSS for the
Arvo font is also included, so the admin bar font matches the admin panel. A
class of 'admin-bar' is added to the body for styling purposes, particularly
to push the rest of the theme down to account for the height of the admin bar.

The admin-bar partial, CSS, and body class are all added only if there is a
current user logged in. The admin-bar is added through the 'public_theme_body'
hook. The styles are added through the CSS queue.

Refs #37.