Adds an AdminForm, with some elements and a decorator, to build forms on the admin side. See #297
Idea is to have an AdminForm class that conforms to the new structure
and CSS of the admin theme, which makes it easier for plugin builders
and us add elements to the appropriate parts of the admin form.

AdminForm has methods for adding elements to the main edit area and to
the save panel. There are elements for common stuff in the save panel
(save, public page link, delete link), hopefully in a way to help plugin
builders add their own elements easily.

Not sure if I've followed naming conventions properly.