Adds options param to javascript queue functions.
Adds an option parameter to `queue_js_file` and `queue_js_string` to allow
for, among other things, wrapping the outputs of those with conditional
comments similar to the CSS queue functions.

Other options can be passed as well, per the Zend Framework documentation for
headScript:

http://framework.zend.com/manual/1.12/en/zend.view.helpers.html#zend.view.helpers.initial.headscript

Updates the test that shall not be named to check for these conditional
comments.