Try to update texts where possible.
The new code matches up existing texts for the same element and just
assigns over the ID instead of creating a new text if one already
existed and we're "overwriting." If there aren't enough old texts,
new ones are still inserted, and if there are too many, the leftovers
are deleted.

This results in much less churn for the common case of one, a few, or no
edits to element texts, all scenarios which currently cause the deletion
and reinsertion of all the texts for the item.