Make Omeka_Db_Select_PublicPermissions really look at the current user.
I'm hesitant about this commit, since it seems odd that we haven't hit this before. Then again, it only came up while testing the combo of UserProfiles and GuestUser

It looks like the situation of having an owner column, and acl with showSelfNotPublic hasn't come up before? (Or at least hasn't been noticed and reported?)

In any case, this rectified a mysql 'not data in prepared statement error' in the combo of UserProfiles and GuestUser. The $select was coming out with the expected OR clause, but no current user id. So far as I can tell, this won't affect the previous behavior, but wanted to be verbose here because it seems. . . curious. And I could always be missing something or causing some unwanted behavior I hadn't thought about.