Make extended_resources always an object.
I didn't see the format documented in the wiki yet, so want to confirm with @jimsafley that this is as it should be, but it looks like same issue as discussed https://github.com/omeka/Omeka/issues/533#issuecomment-17413775

When I extended to include geolocation data on items, extended_resources became an object. When no extensions present, it appeared as an array.