Update files-form.php
By changing order of this form, you can easily add a new file, even if you already have many files associated with the item. You aren't obliger to scroll to the end of the screen to find the "Add files" button.

Going to the bottom of the screen isn't that complicated but when you have a plugin which fires on admin_ites_form_files like Dropbox for example you're sometimes obliged to find this form in the middle of a big screen.