Wrap citation URL in span tag for CSS selection
Because many Omeka sites exist in subdomains or subdirectories, their URLs tend to be rather long. This creates issues for designers as it can break responsive layouts when the URL is so long that it exceeds the width of the container or viewport. 

Wrapping the citation URL in a span allows a designer to apply CSS styles forcing the URL to wrap lines. For example:

`.citation-url{-ms-word-break:break-all;word-break:break-all;word-break:break-word;-webkit-hyphens:auto;-moz-hyphens:auto;hyphens:auto}`