Expand 'display_title' property
The underlying implementation moves from the File record to the
ElementText mixin so it's available for all ElementText-using records,
and the appropriate records have the display_title property added to
their getProperty methods.

The intent is for this to replace the varying incantations of
strip_formatting used (and not used in some places where it should have
been) throughout the views for displaying titles. Essentially all usages
in the codebase of Dublin Core Title being passed to strip_formatting
have been replaced by usages of the display_title property instead.

The specific motivating problem for this change is also fixed by it: a
bug introduced by the previous commit on adding aria-labels which didn't
handle HTML in the titles correctly.