Fix bug with standalone delete-confirm page
The use of the class "delete-confirm" on the confirm form (introduced for
autosave) breaks the page because of pre-existing javascript on that
same class which prevents defaults. This was hidden because it doesn't
occur in the usual "popup" delete-confirm context.

(cherry picked from commit 943d7d973c9c48e70d989018ef463f8edd889147)