Add tag filters for range and collection
I took the filters from the Items Table model and adapted them for use in the Tag Table model. I've tested it on my own Omeka instances, and it successfully filters using range and collection parameters on both the `items/tags` page and the `admin/tags` page. This allows for the generation of tag clouds for Items within a given Collection by adding `&collection=id` to the GET parameters, or within a given range of Item IDs with `range=1,3,5-7` or similar.