Add s3 expiration "chunking"
Current presigned URLs are basically uncacheable as they change for
every request. Instead, treat the expiration as a minimum and tack on
extra time to keep the expiration timestamp steady. The effect is that
the actual expiration time will float between the configured length
(plus one second) and double the configured length.