Update tests to work with PHPUnit 7
Older PHPUnits have issues with PHP 7.4 deprecations. Since these tests
are written for PHPUnit 4, we make judicious use of the
forward-compatibility features, as well as some polyfills of our own for
changed method and class names.

Zend's controller test case code gets somewhat into the internals of
PHPUnit, which aren't as easily handled in this way, so instead of
trying to keep that working with many different versions, this just
ports over simplified versions of the few asserts we actually use and
otherwise totally drops the Zend PHPUnit code. (As well as the unused
Zend_Tool, which has some references to PHPUnit too).