Adding fieldset, aria-live, visually-hidden class
Added fieldsets to edit pages to group input elements, even when new elements are added. Also added an aria-live element to alert screen readers to changed content when adding metadata fields. Last but not least, added a `visually-hidden` class to add content for screen readers that won't be displayed visually on the site, used on browse and edit pages.