Always parse shortcode args as array
The vast majority of shortcode implementations deal with arguments as an
array, but "non-matching" argument strings can be returned as a string
(this includes the common situation where there are no arguments at
all). In old PHP versions the implementations simply assuming the string
was an array would cause it to be transparently converted to one, but in
current PHP versions this instead raises a warning.

Instead of fixing all the implementations, change the shortcode helper
itself to always make the args an array. Empty-string args are now an
empty array, and any other non-matching string is passed as index 0 of
an array rather than as a string directly.

Also, took the opportunity here to improve the comments/annotations on
the argument regex.