PHP 8.1: Use "classic" mysqli error handling
8.1 changes the defaults to have mysqli throw exceptions on errors, but
Zend_Db and our code that depends on it both expect the classic
manual-checking style of error reporting. We set the flag here to opt
back into the previous default.

This fixes, among other things, the redirect to the installer on first
run. More broadly pretty much anything watching for Zend_Db exceptions
would be affected.