Support template mode in navigation link partial.
New links were previously inserted into the navigation via the
javascript using separate markup. This allows us to manage consistent
markup for both new and existing navigation links. The form will include
an empty version of this partial as a template, then the javascript
clones and populates the template with new link data.