Redundantly specify base decimal format for es_419
Zend_Locale doesn't correctly do inheritance and the result is that
locales underneath es_419 have broken number handling. We don't use
these much or at all in the core, but as a band-aid for if plugins do,
this fixes the issue for the es_419 locales (es_CO is our specific one
that users generally use).

(#756)