PHP 8.4: disable E_DEPRECATED reporting
8.4 adds deprecation notices on all methods that use a null default but
don't have a question mark in their typehint, which was/is a very common
pattern. Here we also have the problem that the question mark for
typehints requires PHP 7.1 or greater and the current Omeka Classic
support range goes back further than that.

We'll eventually increase the minimum and make these changes, but will
probably maintain the E_DEPRECATED block at least for a while, as
there's many plugins out in the world with this problem and a low
likelihood of getting fixes.

(see omeka/omeka-s#2233)