Fix password upgrading on verify
We correctly had a call to upgrade a hashed password when trying to
verify if necessary (like when a logged-in user is changing their
password), but the way the update works, the hash would get updated but
the ensuing check would use the old hash and fail.

Fixed by allowing a User to be passed to the hash upgrade function, so
that same object the caller already has will get the updated hash when
an upgrade happens.